ref(preprod): Inline wrapper functions and extract build_changes_map
Remove single-line wrapper functions _find_existing_comment_id and
_save_pr_comment_result, inlining the calls with the hardcoded
"build_distribution" argument directly at the call sites.

Rename _build_changes_map to build_changes_map and move it (along with
its helper _comparison_has_changes) from the status_checks snapshots
tasks module to sentry.preprod.snapshots.utils, since it is now shared
across multiple production modules.
