+++ src/sentry/preprod/snapshots/utils.py
@@ -1,7 +1,10 @@
 from __future__ import annotations
 
 from sentry.preprod.models import PreprodArtifact, PreprodBuildConfiguration
-from sentry.preprod.snapshots.models import PreprodSnapshotComparison
+from sentry.preprod.snapshots.models import (
+    PreprodSnapshotComparison,
+    PreprodSnapshotMetrics,
+)
 
 
 def find_base_snapshot_artifact(
@@ -59,3 +62,37 @@ def find_head_snapshot_artifacts_awaiting_base(
         .select_related("preprodsnapshotmetrics")
         .order_by("-date_added")
     )
+
+
+def _comparison_has_changes(
+    comparison: PreprodSnapshotComparison,
+    fail_on_added: bool = False,
+    fail_on_removed: bool = True,
+) -> bool:
+    return (
+        comparison.images_changed > 0
+        or comparison.images_renamed > 0
+        or (fail_on_added and comparison.images_added > 0)
+        or (fail_on_removed and comparison.images_removed > 0)
+    )
+
+
+def build_changes_map(
+    artifacts: list[PreprodArtifact],
+    snapshot_metrics_map: dict[int, PreprodSnapshotMetrics],
+    comparisons_map: dict[int, PreprodSnapshotComparison],
+    fail_on_added: bool = False,
+    fail_on_removed: bool = True,
+) -> dict[int, bool]:
+    changes_map: dict[int, bool] = {}
+    for artifact in artifacts:
+        metrics = snapshot_metrics_map.get(artifact.id)
+        if not metrics:
+            continue
+        comparison = comparisons_map.get(metrics.id)
+        if not comparison or comparison.state != PreprodSnapshotComparison.State.SUCCESS:
+            continue
+        changes_map[artifact.id] = _comparison_has_changes(
+            comparison, fail_on_added, fail_on_removed
+        )
+    return changes_map
